fix: isActive가 0으로 된 댓글은 이미 삭제된 댓글이라는 예외처리 진행

+++ src/main/java/carrotTeam/carrot/domain/comment/service/CommentService.java
@@ -75,6 +75,9 @@ public class CommentService {
     public void deleteComment(Long comment_id){
 
         Comment foundComment = commentRepository.findById(comment_id).orElseThrow(NotFoundComment::new);
+        if(!foundComment.getIsActive()){
+            throw new NotFoundComment();
+        }
         foundComment.delete();
         commentRepository.save(foundComment);
     }
